class names and containing files renamed to consistent snake_case and CamelCase, with matching file and class names
