+++ widgets/qt_textbox_and_slider.py
@@ -6,7 +6,7 @@ from math import log10
 import qt_custom_sliders
 
 
-class InitializeSliderTextB(QWidget):
+class TextboxAndSlider(QWidget):
     def __init__(self, parent, widget_name, min_range, max_range, data_type, increment, default=0):
         super(QWidget, self).__init__(parent)
